Cleaned up store. Did not have much sense to keep flow related actions and reducers under user rootReducer.
