+++ src/components/Step/Result/components/NodeTypes/OptionNodeType.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {IOptionSummary} from "../../../../../store/user/userTypes";
+import {IOptionSummary} from "../../../../../store/flow/flowTypes";
 import styles from './NodeType.module.scss';
 import {Slider} from "@mui/material";
 import sharedStyles from "../../../../../styles/shared.module.scss";
