+++ src/store/app/appReducer.ts
@@ -1,5 +1,6 @@
 import {createReducer} from "@reduxjs/toolkit";
 import {IAppState, SET_STEP} from "./appTypes";
+import {PURGE} from "redux-persist";
 
 const initialState: IAppState = {
     step: 0
@@ -11,7 +12,8 @@ const appReducer = createReducer(initialState, {
             ...state,
             step: action.payload
         };
-    }
+    },
+    [PURGE]: () => initialState
 })
 
 export default appReducer;
