+++ src/store/store.ts
@@ -6,6 +6,7 @@ import storage from 'redux-persist/lib/storage';
 import userReducer from "./user/userReducer";
 import {TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 import appReducer from "./app/appReducer";
+import flowReducer from "./flow/flowReducer";
 
 const persistConfig = {
     key: 'reducer',
@@ -14,6 +15,7 @@ const persistConfig = {
 
 const reducers = combineReducers({
     user: userReducer,
+    flow: flowReducer,
     app: appReducer
 })
