+++ src/store/user/userAction.ts
@@ -1,5 +1,5 @@
 import { AppDispatch } from "../store";
-import {SET_NAME, START_FLOW, SAVE_FLOW, INewFlow} from "./userTypes";
+import {SET_NAME} from "./userTypes";
 
 export const setUserName = (name: string) => (dispatch: AppDispatch): void => {
     dispatch({
@@ -7,17 +7,3 @@ export const setUserName = (name: string) => (dispatch: AppDispatch): void => {
         payload: name
     })
 };
-
-export const startDecisionFlow = (name: string) => (dispatch: AppDispatch): void => {
-    dispatch({
-        type: START_FLOW,
-        payload: name
-    })
-}
-
-export const saveDecisionFlow = (flow: Partial<INewFlow>) => (dispatch: AppDispatch): void => {
-    dispatch({
-        type: SAVE_FLOW,
-        payload: flow
-    })
-}
